+++ models/Group.go
@@ -8,7 +8,7 @@ import (
 
 // Group 群
 type Group struct {
-	ID             int          `orm:"column(id);auto;pk"`
+	ID             int64        `orm:"column(id);auto;pk"`
 	Name           string       `orm:"size(50)"`                        // 群名称
 	NichName       string       `orm:"size(50)"`                        // 备注群名称
 	Owner          *User        `orm:"rel(fk)"`                         // 群主（一个群只能有一个群主，群主与群的关系是一对一的关系）
